+++ Examples/ActionOpenUrlDialog/Basic Usage.ps1
@@ -22,3 +22,5 @@ New-AdaptiveCard {
         New-CardActionSubmit -Title "Close"
     }
 } | Get-CardResponse
+
+