+++ MvRAdaptiveCards/Public/Application/Send-CardViaTeams.ps1
@@ -3,18 +3,16 @@ function Send-CardViaTeams {
     [CmdletBinding(SupportsShouldProcess = $true)]
     [OutputType([void])]
     param (
-        [Parameter(Mandatory = $true)]
+        [Parameter(Mandatory = $true, ValueFromPipeline = $true)]
         [string]$CardJson,
 
         [Parameter(Mandatory = $false)]
-        [string]$WebhookUrl = 'default'
+        [string]$WebhookUrl = $_MvRACSettings.TeamsWebhook.WebhookUrl
     )
 
     #If the WebhookUrl is set to default, check for module settings
-    if ($WebhookUrl -eq 'default' -and $script:Settings.Teams.WebhookUrl) {
-        $WebhookUrl = $script:Settings.Teams.WebhookUrl
-    }
-    elseif ($WebhookUrl -eq 'default' -and -not $script:Settings.Teams.WebhookUrl) {
+
+    if ([string]::IsNullOrWhiteSpace($WebhookUrl)) {
         throw "No WebhookUrl provided and no default found in module settings."
     }
 
@@ -24,13 +22,13 @@ function Send-CardViaTeams {
         attachments = @(
             @{
                 contentType = "application/vnd.microsoft.card.adaptive"
-                content     = $CardJson
+                content     = ($CardJson | ConvertFrom-Json)
             }
         )
     }
 
     # Send the message to the Teams webhook
     if ($PSCmdlet.ShouldProcess("Send message to Teams webhook at $WebhookUrl")) {
-        Invoke-RestMethod -Uri $WebhookUrl -Method Post -Body ($Payload | ConvertTo-Json -Depth 5) -ContentType "application/json"
+        Invoke-RestMethod -Uri $WebhookUrl -Method Post -Body ($Payload | ConvertTo-Json -Depth $_MaxDepth) -ContentType "application/json"
     }
 }